fix: add validation to null image

+++ app/Http/Controllers/UserController.php
@@ -28,9 +28,11 @@ class UserController extends Controller
             'profile_pic' => ['required', 'image', 'max:5000']
         ]);
 
-        $user_path = storage_path('app/'.$user->profile_pic);
-        if (file_exists($user_path)) {
-            unlink($user_path);
+        if($user->profile_pic){
+            $user_path = storage_path('app/'.$user->profile_pic);
+            if (file_exists($user_path)) {
+                unlink($user_path);
+            }
         }
 
         $path = $request->file('profile_pic');
